pyproject.toml
[tool.poetry]
name = "streaming-data-pipeline"
version = "0.1.0"
description = ""
authors = ["Nicholas <nictann96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
spotipy = "^2.24.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
